moved the states to the component declaration
